Update NZBGet to Trusty, remove unnecessary volumes, cleanup

+++ services/nzbget/start.sh
@@ -2,8 +2,6 @@
 
 set -e
 
-if [ ! -f /data/nzbget.conf ]; then
-  cp /usr/local/share/nzbget/nzbget.conf /data/nzbget.conf
-fi
+[ ! -f /data/nzbget.conf ] && cp /usr/local/share/nzbget/nzbget.conf /data/nzbget.conf
 
 nzbget --configfile /data/nzbget.conf --daemon
